+++ packages/browser-extension/src/modules/service/ado/query.ts
@@ -1,11 +1,16 @@
-export function getRootQuery(rootAreaPath: string, IsDeleted = false) {
+export interface WiqlConfig {
+  rootAreaPath: string;
+  /** @default false */
+  isDeleted?: boolean;
+}
+export function getRootQuery({ rootAreaPath, isDeleted = false }: WiqlConfig) {
   return `
 SELECT
 	[System.Id]
 FROM workitems
 WHERE
 	[System.TeamProject] = @project
-	AND [System.IsDeleted] = ${IsDeleted ? "true" : "false"}
+	AND [System.IsDeleted] = ${isDeleted ? "true" : "false"}
 	AND [System.AreaPath] UNDER '${rootAreaPath}'
 ORDER BY [System.ChangedDate] DESC`.trim();
 }