+++ packages/browser-extension/src/modules/ui/popup.tsx
@@ -31,7 +31,7 @@ export const PopupWindow: React.FC = () => {
   const initialQuery = useRef(localStorage.getItem("last-query") ?? "");
   const [activeQuery, setActiveQuery] = useState(initialQuery.current);
 
-  const setTimestampMessage = useCallback((message: string) => setProgressMessage(`${new Date().toLocaleTimeString()} | ${message}`), []);
+  const setTimestampMessage = useCallback((message: string) => setProgressMessage(`${new Date().toLocaleTimeString()}\xa0 ${message}`), []); // with non-breaking space
 
   const config = useConfigGuard(() => {
     chrome.runtime.openOptionsPage();